+++ statsd/client/__init__.py
@@ -1,0 +1,4 @@
+from __future__ import absolute_import, division, unicode_literals
+
+from .stream import TCPStatsClient, UnixSocketStatsClient  # noqa
+from .udp import StatsClient  # noqa
